requirements.txt
# Core Django
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Environment variables
python-decouple==3.8

# Static files
whitenoise==6.8.2

# Server
gunicorn==23.0.0

# Image processing
Pillow==12.0.0
opencv-python-headless==4.12.0.88

# Data processing
numpy==2.2.6
polars==1.36.1

# Machine Learning
torch==2.5.1
torchvision==0.20.1
ultralytics==8.3.237

# Utilities
requests==2.32.5
PyYAML==6.0.3
